Refactored documentcanvas.tsx
