+++ apps/react/comments/dashboard/inline-comments/open-envoy-demo/components/document/StatusBadge.tsx
@@ -1,0 +1,49 @@
+import { ClockIcon, EyeIcon } from './icons'
+
+interface StatusBadgeProps {
+  status: string
+  type: string
+}
+
+export const StatusBadge = ({ status, type }: StatusBadgeProps) => {
+  const getStatusStyles = () => {
+    switch (type) {
+      case 'review':
+        return 'bg-violet-100 text-violet-700 border-violet-200'
+      case 'pending':
+        return 'bg-gray-100 text-gray-600 border-gray-200'
+      case 'conflict':
+        return 'bg-red-50 text-red-600 border-red-200'
+      case 'dispute':
+        return 'bg-red-50 text-red-600 border-red-200'
+      case 'submit':
+        return 'bg-green-50 text-green-700 border-green-200'
+      default:
+        return 'bg-gray-50 text-gray-600 border-gray-200'
+    }
+  }
+
+  const getIcon = () => {
+    switch (type) {
+      case 'review':
+        return <span className="w-2 h-2 rounded-full bg-violet-500 mr-1.5"></span>
+      case 'pending':
+        return <span className="mr-1"><ClockIcon /></span>
+      case 'conflict':
+        return <span className="w-2 h-2 rounded-full bg-red-500 mr-1.5"></span>
+      case 'dispute':
+        return <span className="w-2 h-2 rounded-full bg-red-500 mr-1.5"></span>
+      case 'submit':
+        return <span className="w-2 h-2 rounded-full bg-green-500 mr-1.5"></span>
+      default:
+        return <span className="mr-1"><EyeIcon /></span>
+    }
+  }
+
+  return (
+    <span className={`inline-flex items-center px-2 py-0.5 rounded text-xs font-medium border ${getStatusStyles()}`}>
+      {getIcon()}
+      {status}
+    </span>
+  )
+}
