+++ apps/react/comments/dashboard/inline-comments/open-envoy-demo/components/document/jobs-data.ts
@@ -1,0 +1,252 @@
+import { Job } from './types'
+
+export const jobsData: Job[] = [
+  {
+    id: 'OE0001',
+    cost: '',
+    comments: 0,
+    ownership: { color: '#10B981', initials: 'E', name: 'Ocean Freight - ...' },
+    due: '...',
+    approver: null,
+    policy: 'AP Eur Matched Appr...',
+    status: 'Matching',
+    statusType: 'default',
+    lineItems: [
+      { id: 'li-1', description: 'Container shipping fee', quantity: 1, unitPrice: 1200, total: 1200 },
+      { id: 'li-2', description: 'Port handling charges', quantity: 2, unitPrice: 150, total: 300 },
+    ]
+  },
+  {
+    id: 'OE0002',
+    cost: '$2,875.93',
+    comments: 32,
+    ownership: { color: '#EC4899', initials: 'R', name: 'Commercial' },
+    due: '-2 days',
+    dueType: 'overdue',
+    approver: { initials: 'LW', color: '#6366F1' },
+    approverTime: '<1d',
+    policy: 'AP US Matched Appr...',
+    status: 'Review & approve',
+    statusType: 'review',
+    lineItems: [
+      { id: 'li-1', description: 'Marketing materials', quantity: 500, unitPrice: 2.5, total: 1250 },
+      { id: 'li-2', description: 'Printing services', quantity: 1, unitPrice: 875.93, total: 875.93 },
+      { id: 'li-3', description: 'Design consultation', quantity: 3, unitPrice: 250, total: 750 },
+    ]
+  },
+  {
+    id: 'OE0003',
+    cost: '$1,729.18',
+    comments: 2,
+    ownership: { color: '#8B5CF6', initials: 'WW', name: 'General AP' },
+    due: '...',
+    approver: null,
+    policy: 'AP Eur Matched Appr...',
+    status: 'Pending',
+    statusType: 'pending',
+    lineItems: [
+      { id: 'li-1', description: 'Office supplies', quantity: 10, unitPrice: 89.50, total: 895 },
+      { id: 'li-2', description: 'Equipment maintenance', quantity: 1, unitPrice: 834.18, total: 834.18 },
+    ]
+  },
+  {
+    id: 'OE0004',
+    cost: '$2,432.82',
+    comments: 0,
+    ownership: { color: '#F59E0B', initials: 'RS', name: 'Packaged Goods' },
+    due: '-10 days',
+    dueType: 'overdue',
+    approver: { initials: 'LW', color: '#6366F1' },
+    approverTime: '<1d',
+    policy: 'AP Eur Matched Appr...',
+    status: 'Review & approve',
+    statusType: 'review',
+    lineItems: [
+      { id: 'li-1', description: 'Packaging materials', quantity: 1000, unitPrice: 1.25, total: 1250 },
+      { id: 'li-2', description: 'Label printing', quantity: 5000, unitPrice: 0.15, total: 750 },
+      { id: 'li-3', description: 'Shipping boxes', quantity: 100, unitPrice: 4.33, total: 432.82 },
+    ]
+  },
+  {
+    id: 'OE0005',
+    cost: '$16,431.76',
+    comments: 0,
+    ownership: { color: '#8B5CF6', initials: 'MJ', name: 'Ocean Freight - ...' },
+    due: 'Today',
+    dueType: 'today',
+    approver: { initials: 'LW', color: '#6366F1' },
+    approverTime: '3d',
+    policy: 'Metals & Timber',
+    status: 'Review & approve',
+    statusType: 'review',
+    lineItems: [
+      { id: 'li-1', description: 'Steel beams shipment', quantity: 50, unitPrice: 250, total: 12500 },
+      { id: 'li-2', description: 'Insurance premium', quantity: 1, unitPrice: 2431.76, total: 2431.76 },
+      { id: 'li-3', description: 'Customs clearance', quantity: 1, unitPrice: 1500, total: 1500 },
+    ]
+  },
+  {
+    id: 'OE0006',
+    cost: '$115,472.51',
+    comments: 5,
+    ownership: { color: '#10B981', initials: 'SC', name: '3rd Party Suppli...' },
+    due: '12 days',
+    dueType: 'warning',
+    approver: { initials: 'LW', color: '#6366F1' },
+    policy: 'Vegetable protein ap...',
+    status: 'Conflict',
+    statusType: 'conflict',
+    lineItems: [
+      { id: 'li-1', description: 'Bulk protein powder', quantity: 10000, unitPrice: 8.50, total: 85000 },
+      { id: 'li-2', description: 'Quality testing', quantity: 1, unitPrice: 5472.51, total: 5472.51 },
+      { id: 'li-3', description: 'Expedited shipping', quantity: 1, unitPrice: 25000, total: 25000 },
+    ]
+  },
+  {
+    id: 'OE0007',
+    cost: '$5,820.91',
+    comments: 0,
+    ownership: { color: '#3B82F6', initials: 'JK', name: 'Commercial' },
+    due: '47 days',
+    approver: { initials: 'LW', color: '#6366F1' },
+    policy: 'Vegetable protein ap...',
+    status: 'Conflict',
+    statusType: 'conflict',
+    lineItems: [
+      { id: 'li-1', description: 'Trade show booth rental', quantity: 1, unitPrice: 3500, total: 3500 },
+      { id: 'li-2', description: 'Display materials', quantity: 1, unitPrice: 2320.91, total: 2320.91 },
+    ]
+  },
+  {
+    id: 'OE0008',
+    cost: '$3,711.83',
+    comments: 0,
+    ownership: { color: '#6366F1', initials: 'EL', name: 'General AP' },
+    due: '8 days',
+    dueType: 'warning',
+    approver: { initials: 'AK', color: '#F59E0B' },
+    approverTime: '1d',
+    policy: 'Metals & Timber',
+    status: 'Submit for approval',
+    statusType: 'submit',
+    lineItems: [
+      { id: 'li-1', description: 'Timber processing fee', quantity: 1, unitPrice: 2500, total: 2500 },
+      { id: 'li-2', description: 'Quality inspection', quantity: 1, unitPrice: 711.83, total: 711.83 },
+      { id: 'li-3', description: 'Documentation fee', quantity: 1, unitPrice: 500, total: 500 },
+    ]
+  },
+  {
+    id: 'OE0009',
+    cost: '$8,348.07',
+    comments: 0,
+    ownership: { color: '#3B82F6', initials: 'JK', name: 'N. American Over...' },
+    due: 'Paid',
+    dueType: 'paid',
+    approver: { initials: 'LW', color: '#6366F1' },
+    policy: 'AP Eur Matched Appr...',
+    status: 'Dispute',
+    statusType: 'dispute',
+    lineItems: [
+      { id: 'li-1', description: 'Ground transportation', quantity: 15, unitPrice: 450, total: 6750 },
+      { id: 'li-2', description: 'Fuel surcharge', quantity: 1, unitPrice: 1598.07, total: 1598.07 },
+    ]
+  },
+  {
+    id: 'OE0010',
+    cost: '$11,096.22',
+    comments: 3,
+    ownership: { color: '#8B5CF6', initials: 'WW', name: 'Trucking' },
+    due: '11 days',
+    approver: { initials: 'LW', color: '#6366F1' },
+    policy: 'AP Mex Matched App...',
+    status: 'Verifying',
+    statusType: 'default',
+    lineItems: [
+      { id: 'li-1', description: 'Long haul trucking', quantity: 3, unitPrice: 3200, total: 9600 },
+      { id: 'li-2', description: 'Loading/unloading', quantity: 6, unitPrice: 249.37, total: 1496.22 },
+    ]
+  },
+  {
+    id: 'OE0011',
+    cost: '$85,682.06',
+    comments: 0,
+    ownership: { color: '#3B82F6', initials: 'JK', name: 'Journal entries' },
+    due: 'Paid',
+    dueType: 'paid',
+    approver: { initials: 'LW', color: '#F59E0B', hasIndicator: true },
+    approverTime: '6d',
+    policy: 'AP Mex Matched App...',
+    status: 'Approvals',
+    statusType: 'default',
+    lineItems: [
+      { id: 'li-1', description: 'Q4 inventory adjustment', quantity: 1, unitPrice: 45000, total: 45000 },
+      { id: 'li-2', description: 'Depreciation entries', quantity: 1, unitPrice: 25682.06, total: 25682.06 },
+      { id: 'li-3', description: 'Accrual reversal', quantity: 1, unitPrice: 15000, total: 15000 },
+    ]
+  },
+  {
+    id: 'OE0012',
+    cost: '$7,026.34',
+    comments: 0,
+    ownership: { color: '#10B981', initials: 'GB', name: 'Commercial' },
+    due: '32 days',
+    approver: { initials: 'MN', color: '#EC4899' },
+    approverTime: '8d',
+    policy: 'Metals & Timber',
+    status: 'Approvals',
+    statusType: 'default',
+    lineItems: [
+      { id: 'li-1', description: 'Client entertainment', quantity: 1, unitPrice: 3500, total: 3500 },
+      { id: 'li-2', description: 'Travel expenses', quantity: 1, unitPrice: 2526.34, total: 2526.34 },
+      { id: 'li-3', description: 'Conference registration', quantity: 2, unitPrice: 500, total: 1000 },
+    ]
+  },
+  {
+    id: 'OE0013',
+    cost: '$7,026.34',
+    comments: 27,
+    ownership: { color: '#10B981', initials: 'GB', name: 'Commercial' },
+    due: '32 days',
+    approver: { initials: 'EA', color: '#6366F1' },
+    approverTime: '2d',
+    policy: 'Metals & Timber',
+    status: 'Review & approve',
+    statusType: 'review',
+    lineItems: [
+      { id: 'li-1', description: 'Advertising spend', quantity: 1, unitPrice: 5000, total: 5000 },
+      { id: 'li-2', description: 'Media placement', quantity: 1, unitPrice: 2026.34, total: 2026.34 },
+    ]
+  },
+  {
+    id: 'OE0014',
+    cost: '$7,026.34',
+    comments: 9,
+    ownership: { color: '#10B981', initials: 'GB', name: 'Commercial' },
+    due: '32 days',
+    approver: { initials: 'TM', color: '#F59E0B' },
+    approverTime: '11d',
+    policy: 'AP Eur Matched Appr...',
+    status: 'Approvals',
+    statusType: 'default',
+    lineItems: [
+      { id: 'li-1', description: 'Consulting services', quantity: 20, unitPrice: 275, total: 5500 },
+      { id: 'li-2', description: 'Project expenses', quantity: 1, unitPrice: 1526.34, total: 1526.34 },
+    ]
+  },
+  {
+    id: 'OE0015',
+    cost: '$7,026.34',
+    comments: 0,
+    ownership: { color: '#10B981', initials: 'GB', name: 'Commercial' },
+    due: '32 days',
+    approver: { initials: 'MS', color: '#6366F1' },
+    approverTime: '2d',
+    policy: 'AP Mex Matched App...',
+    status: 'Approvals',
+    statusType: 'default',
+    lineItems: [
+      { id: 'li-1', description: 'Software licenses', quantity: 10, unitPrice: 450, total: 4500 },
+      { id: 'li-2', description: 'Support contract', quantity: 1, unitPrice: 2526.34, total: 2526.34 },
+    ]
+  },
+]
