+++ apps/react/comments/dashboard/inline-comments/open-envoy-demo/components/document/types.ts
@@ -1,0 +1,22 @@
+export interface JobLineItem {
+  id: string
+  description: string
+  quantity: number
+  unitPrice: number
+  total: number
+}
+
+export interface Job {
+  id: string
+  cost: string
+  comments: number
+  ownership: { color: string; initials: string; name: string }
+  due: string
+  dueType?: string
+  approver: { initials: string; color: string; hasIndicator?: boolean } | null
+  approverTime?: string
+  policy: string
+  status: string
+  statusType: string
+  lineItems: JobLineItem[]
+}
