+++ apps/react/comments/dashboard/inline-comments/open-envoy-demo/components/sidebar/sidebar.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import { VeltNotificationsTool, useNotificationUtils } from '@veltdev/react';
+import { VeltNotificationsTool, useNotificationUtils, useUnreadNotificationsCount } from '@veltdev/react';
 
 // Sidebar icon assets
 const imgOpenEnvoyWordmarkSmallUsePositive1 = "/icons/sidebar/open-envoy-logo.svg";
@@ -82,10 +82,14 @@ function SectionHeader({ label }: SectionHeaderProps) {
 
 export default function Sidebar() {
   const notificationUtils = useNotificationUtils();
+  const unreadCount = useUnreadNotificationsCount();
 
   const handleNotificationsClick = () => {
     notificationUtils?.openNotificationsPanel();
   };
+
+  // Get the count for "For You" tab (default), fallback to 0
+  const notificationCount = unreadCount?.forYou ?? 0;
 
   return (
     <div className="flex h-full">
@@ -130,19 +134,30 @@ export default function Sidebar() {
             label="Search"
           />
           {/* Notifications - entire row clickable */}
-          <button
-            onClick={handleNotificationsClick}
-            className="w-full h-8 flex items-center gap-3 px-2 py-1 rounded-[3px] hover:bg-[#F2F4FF]/40 transition-colors duration-150"
-          >
-            <img src={imgNotifications} alt="" className="w-4 h-4 flex-shrink-0 block" />
-            <span
-              className="flex-1 text-left text-sm leading-[21px] text-[#4A4947] tracking-[-0.105px]"
-              style={{ fontFamily: 'Inter, sans-serif' }}
+          <div className="relative">
+            <button
+              onClick={handleNotificationsClick}
+              className="w-full h-8 flex items-center gap-3 px-2 py-1 rounded-[3px] hover:bg-[#F2F4FF]/40 transition-colors duration-150"
             >
-              Notifications
-            </span>
-            <VeltNotificationsTool shadowDom={false} />
-          </button>
+              <img src={imgNotifications} alt="" className="w-4 h-4 flex-shrink-0 block" />
+              <span
+                className="flex-1 text-left text-sm leading-[21px] text-[#4A4947] tracking-[-0.105px]"
+                style={{ fontFamily: 'Inter, sans-serif' }}
+              >
+                Notifications
+              </span>
+              <span
+                className="flex items-center justify-center min-w-[18px] h-[18px] px-[5px] bg-[#BD323C] rounded-[5px] text-white text-xs font-medium leading-[18px] tracking-[0.25px]"
+                style={{ fontFamily: 'Inter, sans-serif' }}
+              >
+                {notificationCount}
+              </span>
+            </button>
+            {/* VeltNotificationsTool positioned at end of row for panel placement */}
+            <div className="absolute right-0 top-0 h-full flex items-center opacity-0 pointer-events-none">
+              <VeltNotificationsTool />
+            </div>
+          </div>
           <NavItem
             icon={imgHome}
             label="Home"
